Added support for using more than one rtp proxy, in which case traffic will be distributed evenly among them. In addition, each such proxy can be assigned a weight, which will specify which share of the traffic should be placed to this particular proxy.
Introduce failover mechanism, so that if SER detects that one
of many proxies is no longer available it temporarily decreases
its weight to 0, so that no traffic will be assigned to it.
Such "disabled" proxies are periodically checked to see if they
are back to normal in which case respective weight is restored
resulting in traffic being sent to that proxy again.

Those features can be enabled by specifying more than one "URI"
in the rtpproxy_sock parameter, optionally followed by the weight,
which if absent is assumed to be 1, for example:

rtpproxy_sock="unix:/foo/bar=4 udp:1.2.3.4:3456=3 udp:5.6.7.8:5432=1"