Properly handle integer columns in the subscriber table when generating the AVPs specified by the auth_db/load_credentials module parameter. This fixes crahses caused by invalid memory access when a non-string column was specified to be loaded in load_credentials.
With this patch, string and integer table columns are properly handled and
will create an AVP with a string respectively an integer value depending
on the type of the column in the subscriber table that is loaded.
Other column types are currently not handled and will generate an error
message in the logs if used, however if need arises they can be handled by
converting them to a string and creating an AVP with a string value.

(patch ported from the stable branch)


git-svn-id: https://openser.svn.sourceforge.net/svnroot/openser/trunk@826 689a6050-402a-0410-94f2-e92a70836424