Added parameter to dlg_request_uas and dlg_response_uac.
This parameter specifies if it is target refresh request/response. This is
needed because TM module is not able to determine target refresh methods for
all dialog types.  Only INVITE method was used as target refresh, but this is
error for non-INVITE dialogs (SUBSCRIBE/NOTIFY, ...).

If you want to behave like previous versions use TARGET_REFRESH_UNKNOWN as
value but be aware of possible problems.