backport to 2.0
Allow to compare the snd_ip and to_ip with the AVP value. (Was allowed for ports, not ips.)
For performance purist: I did not measure which code is faster; whether finding the AVP in the list or using DNS name resolution. Using AVP (track and) domain prefix makes the search faster, when AVP is in fallback domains, e.g. for global AVP.

Following code now allowed:
onsend_route{
  if (isflagset(GATEWAY)) {
    if (to_ip!=$g.gw_ip) {
          log(1,"Bad gateway IP");
          drop;
        }
  }
}

P.S. Regarding latest discussion on bugs/features.... it can be seen as "bug" because port was allowed to be compared to AVP, but ip not. I'm sorry for any inconvinience. ;-)