Various db api changes needed after migration to libsrdb1
Database fixes done by kam_to_sr.sh script

This patch makes the module compatible with the changes done in the
database library of kamailio/ser when it was converted to shared library
libsrdb1. These changes are mostly needed to avoid conflicts with the
symbols defined in the newer version of the library libsrdb2.

List of changes:
 * Rename db_con_t to db1_con_t
 * Rename database field type names from DB_* to DB1_*