Define flag FL_REQ_UPSTREAM in acc module.
This flag was originally defined in file msg_parser.h. But the
sip-router core does not contain this definition and it does not make
much sense to add it there because the acc module is the only module
using it and later, if we merge acc implementations from ser and
kamailio, this flag will be probably gone.

File sip_router/msg_parser.h contains a warning about the flag being
defined here, this is to make sure we don't create another flag with
conflicting value.