Conversion of flags from kamailio to ser flag interface
The original (kamailio) version of acc module accesed the flags in
sip_msg structure directly through a few macros declared in this
module. In SER we introduced the concept of named flags and created a
set of functions that can be used to access and modify the flags.

First of all we need to replace calls to flag_idx2mask from mod_init
function, this function does not exist in the sip-router core because
all flag related functions there get the index of the flag as parameter.
This means that we do not have to convert values of flag-related
modparams in this module, all we have to do is check whether they have
a valid value. This is done using flag_is_in_range. But because the
module uses -1 to denote that a particular flag is not set, we need
only need to call that function if the flag is not set to -1.

After that we need to update is_acc_flag_set and reset_acc_flag macros
to call functions from flags.h, instead of modifying the flags
directly. But again, the functions from flags.h should only be called
if the particular flag is not set to -1.

If we also wanted to add support for named flags then we would need to
update the definition of modparams holding the value of flags in this
module. This is not done in this patch. We can do it later after we
merge acc modules from ser and kamailio.