Add logging facility parameter to LOG_.
The recent changes in dprint.h introduced a new parameter to LOG_ macro
which can be used to specify other than the default logging facility
configured through the config framework.

All calls to LOG_ need to be updated to have DEFAULT_FACILITY as the
first parameter. Other logging macros are unaffected.