A comment warning about flag FL_REQ_UPSTREAM.
This patch adds a comment warning about flag FL_REQ_UPSTREAM which is
defined in kamailio module acc and uses the value (1<<11) in sip message
flags. This flags was originally defined in this file, but since the
kamailio acc module is the only module setting and using the flag.

I think it would make more sense to define it in the module itself
before we merge ser and kamailio acc implementations (then the flag
probably won't be needed anymore because the test for the direction of
the message is trivial and there is no need to set the flag in the sip
message, especially if it is not being used by any other module).

Signed-off-by: Jan Janak <jan@iptel.org>