Function load_tm_api added
There are many modules that need to access the API of tm module and they all
have to go through the same step of commands to bind the API exported by
the tm module. It makes more sense to create one function which does it all
and make that function available to all other modules using tm.

Signed-off-by: Jan Janak <jan@iptel.org>