Update parameters of append_branch for the sip router core.
The function append_branch in the sip-router core accepts pointers and
lengths of strings parameters separately, this is useful when you don't
know in advance whether you are dealing with str type of strings or
zero terminated strings.

Also there currently is no parameter for branch flags, this is to be
added later and then we will updated the patch.