Switch to sip-router timer api.
This removes the function register_timer_process and leaves only
register_timer. Function register_timer_process is missing in
sip-router and as far as I can tell, it can be replaced by
register_timer.

It also removes the associated timer_mode modparam which, provided
that we can use register_timer instead of register_timer_process,
should not be needed anymore.