Move MD5File into cfgutils module.
The function MD5File was originally defined in kamailio/md5utils.c, but
since there is only one module using it (cfgutils), it makes sense to
store the function directly in the module, as opposed to the shared
library likcore or the core itself. It can be always moved if it ever
happends that it will be needed by another module or the core itself.