Use sip-router style of calling do_action.
Before calling do_action function from the sip router core, one has to
first create and initialize a context variable of type run_act_ctx.
Pointer to this variable is then passed as the first parameter to
do_action.

Also, it is safer to zero the whole 'action' structure using a memset
than setting all its attributes explicitly, you never know when
somebody adds a new attribute to the structure and then it might be
used unitialized.

Finally, attribute 'elem' in 'action' structure was renamed to
'val in ser.