Use sip-router style of calling do_action.
Before calling do_action function from the sip router core, one has to
first create and initialize a context variable of type run_act_ctx.
Pointer to this variable is then passed as the first parameter to
do_action.

Attribute 'elem' in 'action' structure was renamed to 'val in ser.