Update calls to search_first_avp to sip-router convention.
The sip-router core contains a structure called search_state which is
can be used to keep the state of the avp search. A pointer to a variable
of this type is given to search_first_avp. To restart the search one
can call search_next_avp, giving it the search_state variable used with
search_first_avp as a parameter.

Also there is few more types defined in usr_avp.h in sip-router, such
as avp_list_t and avp_t.