Kamailio compatibility: script flags
This patch adds support for script flags identical to those present in
the kamailio core. Script flags is a global set of flags which preserve
their value for the duration of script processing. Script flags are kept
in a global variable in the sip-router core.

There are several functions that can be used to manipulate the value
of script flags:
 * setflagsval - This function sets the value of _all_ script flags
   at once. The new value of all the flags must be combined in the
   single function parameter.

 * setsflag - This function sets one particular script flag to 1. The
   function gets the index of the flag (counting from 0) as a
   parameter.

 * resetsflag - This function sets one particular script flag to 0. The
   function gets the index of the flag (counting from 0) as a parameter.

 * issflagset - Test the value of a script flag. Returns 1 if the flag
   is set and -1 otherwise.

 * getsflags - Returns the value of all script flags combined into a
   single variable of type flag_t. This function can be used to make a
   backup copy of script flags.