Fix perl compilation problem.
Adding the output of the perl subcommand enclosed in `` causes the value
of saved DEFS to be different than the value in the makefile and this
results in a never-ending compilation loop. Instead of that we save the
output of the perl subcommand in PERLCCOPTS and then add the value of
PERLCCOPTS to DEFS.

Suggested by Andrei.