core: remove VQ_MALLOC memory manager
attached patch removes the VQ_MALLOC memory manager from the
sip-router core.  As discussed some time ago, this memory manager
doesn't provide a realloc implementation, and is thus not really
usable for the modules, as some of them needs this. We removed
this in kamailio in the last year, so far nobody missed it.

Signed-off-by: Andrei Pelinescu-Onciul <andrei@iptel.org>