Copy parameter hooks into the shm mem copy of the event structure.
When copying an event structure into the shared memory, update also
the value of dialog event parameter hooks and make them point to
the parameters duplicated in shared memory. This is needed so that
the macro EVENT_DIALOG_SLA can still be used on the shared memory copy
of the event (the macro tests if the sla parameter hooks is not NULL).