New functions that can be used to search for a header with given type
This patch adds two new functions to the sip-router core. Function
get_hdr can be used to find the first occurrence of header with a given
type in a sip message if the particular header type does not have its
own hook in the sip_msg data structure.

The second function next_sibling_hdr can be used to find the next
occurence of header with the same type as the header in parameter
in the sip message.

Signed-off-by: Jan Janak <jan@iptel.org>