Switch to the openser module interface version
We need to switch the tls module to the openser version of module
interface in order to make all the tls related pseudo-variables
available.

This comes with a price, we won't be able to use the TLS rpc
commands temporarily because they can only be exported through
the SER version of the module interface. But there are only two
functions available through the rpc interface anyway and not
having them harms less than not having the pseudo-variables.