makefile: added CFG_NAME and SCR_NAME
CFG_NAME and SCR_NAME are used now to hold the name or prefix for
the config and scripts _in_ the repository. All this configs and
scripts will be installed using MAIN_NAME
(e.g.  $(CFG_NAME)-basic.cfg is installed into
       etc/$(MAIN_NAME)/$(MAIN_NAME)-basic.cfg).
This was needed to allow installing using different names, without
having to revert the recent mass re-naming of man, config and
scripts.