domain: Add internal function is_domain_local
Add internal function is_domain_local that can be used to determine if
the domain name given in parameter is on the list of local domains. The
comparison of the domain name is always case insensitive.

There already is is_local function that does something similar, so we
can reuse the code from that function. is_domain_local takes a pointer
to str as parameter and this function is then called by is_local which
takes all the usual parameters that function which can be called from
the script take.

Finally, is_domain_local is exported through the internal API of the
domain module.

Function is_domain_local is now located in file domain.c and because of
db_get_did was also moved to file domain.c