xmlrpc(s): option for replacing double lf  with crlf
Added "double_lf_to_crlf" config parameter. When set double LFs in
the input xmlrpc strings are replaced with CR LF. This is needed
to work around some xmlrpc buggy clients that don't escape CR
('\r' should be escaped to &#xD; ). When using one of those buggy
clients, use double LF as an escape for CR LF and turn this
 option on.Turning this option on makes it also more compatible with
kamailio tools (kamailio mi xmlrpc always converts double LF to CR
LF).
Note that when double_lf_to_crlf is on it's not possible to send
a double lf.

Here is what the xmlrpc modules sees for various combinations of
CR and LF in the input, when CR is _not_ escaped:
         0        1      (double_lf_to_crlf)
\n   -> \r       \r
\r   -> \r       \r
\r\r -> \n\n     \r\n
\n\n -> \n\n     \r\n
\r\n -> \n !!!   \n !!!
\n\r -> \n\n     \r\n

Note that when double_lf_to_crlf is 0, the xmlrpc module behaves
correctly according to the xml specs.